feat: commit files from `assets` even if in they are in `.gitignore`
BREAKING CHANGE: the `.gitignore` file is now ignored when adding files to the release commit

Each files matching a glob in `assets` will now be included in the release commit. The `assets` option has now to be configured to match exactly the the files that have to be commited, independently of the `.gitignore` file.