fix: glob `assets` before running `git ls-files`
The command `git ls-files -m -o` seems to behave differently based on the Git version used. In addition it fails on certain Git versions when `.git` or `.gitmodules` are included in the project dependencies.
This change glob the `assets` first, avoiding to run `git ls-files -m -o` on the entire repo.